refactor: Controller, RequestMapper 의존성 정리
- RequestMapperImpl (기존 RequestMapper)클래스 catalina로 이동
- RequestMapper interface 생성: coyote
- Controller interface 이동: catalina -> coyote
