+++ tomcat/src/main/java/org/apache/catalina/connector/Connector.java
@@ -1,5 +1,6 @@
 package org.apache.catalina.connector;
 
+import org.apache.catalina.controller.RequestMapperImpl;
 import org.apache.coyote.http11.Http11Processor;
 import org.slf4j.Logger;
 import org.slf4j.LoggerFactory;
@@ -66,7 +67,7 @@ public class Connector implements Runnable {
         if (connection == null) {
             return;
         }
-        var processor = new Http11Processor(connection);
+        var processor = new Http11Processor(connection, RequestMapperImpl.getInstance());
         new Thread(processor).start();
     }
