+++ tomcat/src/main/java/org/apache/coyote/http11/Controller.java
@@ -1,4 +1,4 @@
-package org.apache.catalina.controller;
+package org.apache.coyote.http11;
 
 import org.apache.coyote.http11.request.HttpRequest;
 import org.apache.coyote.http11.response.HttpResponse;
@@ -6,5 +6,6 @@ import org.apache.coyote.http11.response.HttpResponse;
 public interface Controller {
 
     boolean isMappedController(HttpRequest request);
+
     void service(HttpRequest request, HttpResponse response) throws Exception;
 }
